Update help.js
In the module.exports: 

permissions = [],

is defined and by default the permission check uses let permissions = command.permission which is not used in the add or add-economy command. This leads to all commands being shown for all users irrespective of their actual permissions. Changed permissions to commandPermissions to be prevent overlap with existing variable and to be less ambiguous & to function as intended.